+++ R/Optimizer.R
@@ -11,6 +11,11 @@
 #' of the [OptimInstance] at the end in order to store the best point  and its
 #' estimated performance vector.
 #'
+#' @section Progress Bars:
+#' `$optimize()` supports progress bars via the package \CRANpkg{progressr}
+#' combined with a [Terminator]. Simply wrap the function in
+#' `progressr::with_progress()` to enable them. We recommend to use package
+#' \CRANpkg{progress} as backend; enable with `progressr::handlers("progress")`.
 #' @export
 Optimizer = R6Class("Optimizer",
   public = list(